Simplify proxying -- don't interpret the request, just forward.
We just set query parameters to exactly req.query now. This prevents the
server from having to know anything about the request.

We also just relay the response body with the same response code instead
of trying to error handle on non-200 codes, since we don't know how
keybase will respond and want to duplicate it.