Add React-ful invite page at GET /invite
Add a new endpoint, /invite, which serves the UI for viewing received
invites.

The invite page is a single React component, InviteClient, which is
responsible for fetching the invite data and decrypting it.

Unfortunately, I wasn't able to reuse the Message class from the
standard email client, because Message is responsible for decryption,
and it assumes things like having a p3skb key in localStorage. Possible
candidate for refactoring in the future.