Fix logins by using URL escape characters in login_session query param
The problem with logins was that getsalt.json returns a login_session,
which is a base64 string. That string is sent back in a JSON body.
login.json requires us to send that string back, however the base64
encoding was not URL safe, which resulted in the bad session issue.

Initially I tried using the URLSafeBase64 module to fix this, but it
turns out that Keybase doesn't use the url safe base64 convention.
I had to manually replace the non URL safe characters with the URL
escape characters.

The reason this probably regressed after migrating to XHR is because
I think XHR doesn't automatically escape invalid URL characters for you,
while the request module did.