Store a single full thread ID, check it at render time in ThreadSnippet.
This is better than storing the information directly in the
MessageStore's thread objects for a couple of reasons:

1. It agrees more fluently with the invariants of the system, wherein
only one thread can be open at a time (so it makes sense to simply hold
on to that thread's ID, if any, in a centralized location, rather than
separate booleans per thread).

2. Extra information stored in `_threads` in the `MessageStore` is lost
when a refresh happens, since all those objects get replaced.