Remove header.ejs from mail.ejs; only use Header.react in /mail.
Previously, we were briefly loading header.ejs while the
EmailClient was in the process of loading. This required us to
pass the email variable from the server, which it no longer needs
access to. So I removed the template-based header. This doesn't
change the main functionality of the page since the actual header
during the main operation of the app is generated by the Header
React component, loaded by EmailClient.react.js.