Add notification of update availability into the header on Electron.
(Only in Electron...)
MessageStore now checks regularly for the latest version number, and
exposes a method to get it. The Header react component gets the latest
version number, and compares it to the app's current version number. It
displays a link in the Header which redirects the browser to the
download section of the website.

Modified the index page to have an anchor at #download to jump to.